x86: don't use REP_GOOD or ERMS for user memory clearing
The modern target to use is FSRS (Fast Short REP STOS), and the other
cases should only be used for bigger areas (ie mainly things like page
clearing).

Note! This changes the conditional for the inlining from FSRM ("fast
movs") to the proper FSRS ("fast stos").  We should check if there are
microarchitectures that have a good 'rep stosb' that do not set that bit
(ie particularly some AMD).

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>